fix(mongoose): fix Document and Subdocument .parent()
`Document.prototype.parent()` exists and returns the parent Document if `this` is a
Subdocument or a populated document, and `undefined` otherwise:
https://mongoosejs.com/docs/api/document.html#document_Document-parent.

Instead, this was only defined on `Embedded.prototype` and missing on `Document.prototype`
(where it can return `undefined | Document`) and on `Subdocument.prototype` (where it
always returns its parent `Document`).

This all constrasts with mongoose 4, which only defined `.prototype.parent()` on
`Subdocument` and `EmbeddedDocument`:
https://mongoosejs.com/docs/4.x/docs/api.html#types_subdocument_Subdocument-parent. So it
seems these types just didn't get updated for mongoose 5.