Update/auth0-lock: Add hidden type and storage property support for signup fields
According to https://auth0.com/docs/libraries/lock/lock-configuration#hidden-field there's a hidden field type available.
This type is currently missing from the types.
According to https://auth0.com/docs/libraries/lock/lock-configuration#text-fields there's also a storage property which can be set to "root". It's only documented on the text field type but I'm not sure if this is accurate since the limitation on it's usage base on the field name and not necessarily on the field type. 
Hence I added the storage property to all types.