ActionCable: Support contextual methods.
This is a follow-up for #52421. Returning an extended type caused the
contextual methods from `Channel` to be dropped.

This commit brings back support to reference methods through `this`.

I've added a test here and also tested against the codebase showing the
original problem in: https://github.com/DefinitelyTyped/DefinitelyTyped/pull/52421#issuecomment-828503478