fix(stopword): correct literal union value
There was a typo in one of the language code exports in previous update.
To prevent this in feature and ease usage in TypeScript, this moves all
language code into exported union types and changes the export definitio
for a module (backward compatible).
All language code are now just keyed (indexed) type on the module
definition.

Thanks!

/cc @spicemix @fergiemcdowall

Fixes fergiemcdowall/stopword#179