style should allow null
null in style have meaning not like undefined

Style for the selected features. By default the default edit style is used (see module:ol/style). Set to null if this interaction should not apply any style changes for selected features. If set to a falsey value, the selected feature's style will not change.