[trusted-types] Move core type definitions into 'trusted-types/lib'
The 'trusted-types' entrypoint attaches the Trusted Types type
definitions to the global scope. As a native DOM API, Trusted Types
should indeed be available globally, but this makes it unattractive for
library authors to adopt the type definitions, as this pollutes the
global scope for all its users. Even worse, when Trusted Types become
part of lib.dom.d.ts the duplicate type definitions will cause a
compilation error that is very tricky to address, both at the
application and library level.

To circumvent this, introduce a new 'trusted-types/lib' entrypoint
intended specifically for libraries that want to adopt the types. The
core type definitions are moved into this entrypoint but then
re-exported into the global scope in the 'trusted-types' endpoint, which
application authors can continue to use. The main caveat is that
libraries will need to explicitly import all the types that they use,
and cast the window object as TrustedTypesWindow to access its Trusted
Types properties.