Updated gulp-sass definition to v5
The gulp-sass@v5 has breaking change in the main export. Now the module requires the compiler parameter.
The compiler is another module which gulp-sass wraps. Because the compiler module can have no definition (or can have wrong/outdate one), the compiler param is typed as any.
This is enough to address the breaking change, keep the typescript compiler happy and the developer experience smooth (no need to precisely type the compiler param and track such dependency).