Missing Type definition from version-1.3.0
In latest update (1.3.0 ) property "retainEndDateOnFirstSelection" was added to stop end date to change to change when star date is change, retainEndDateOnFirstSelection is by default false and when set to true, it retains the end date on every start date change.... To be noted, when moveRangeOnFirstSelection  is set to true,  retainEndDateOnFirstSelection={true} will be ignored. And this property was missing in type definition.