Proposal to remove `any` and replace with `unknown`
Anything that comes off the wire is essentially `unknown`.  My feeling is that `any` was used before `unknown` was available.  This is definitely a breaking change, but I'd like to propose it to stop developers passing properties on body without thinking of validation.

This has been around since TS 3.0 and is described as:

TypeScript 3.0 introduces a new top type unknown. unknown is the type-safe counterpart of any. Anything is assignable to unknown, but unknown isn’t assignable to anything but itself and any without a type assertion or a control flow based narrowing. Likewise, no operations are permitted on an unknown without first asserting or narrowing to a more specific type.