Fix @ckeditor/ckeditor5-utils Options target
For the `Options` interface, modify the `target` and `limiter` properties to accept whatever the constructor parameters
of `Rect` are, cause these two properties are used unmodified to create a new `Rect`.

For the Rect constructor, add `DOMRect` which is the standard version of ClientRect, the window object, and the expected
type for the record.