Add 'rotation' and 'direction' to LayoutAxis
In the underlying JavaScript code, plots such as the 'scatterpolar' chart allow for the rotation of the angular axis as well as changing the direction with which the data is drawn (clockwise / counter-clockwise). I propose updating the 'LayoutAxis' interface to allow TypeScript code to interface with these properties. The changes have been tested on my end in a client application and shown to work correctly.