Move Buffer to an export with a global alias
Currently Buffer is declared as a global and then exported from inside
the 'buffer' module. This is disallowed when there is a merge because
its declarations start coming from a file outside buffer.d.ts.

However, this prevents Buffer merges from specifying what class they
extend because the name is now private. That means pouchdb-core needs to
change.