Making arguments optional
With never versions of the AWS lambda i noticed that VSCode is really wanting me to input that callback method when bundling and using function by import between lambdas.
and since the API states that perhaps the 2 last arguments are often omitted and not needed i propose that we make those arguments to the Handler type optional.

guess maby this is auto generated from something but i'll post the proposal and see if it goes through! best regards!