[mdast]: allow `null` wherever `undefined` is allowed
Given that the actual code generating mdast (such as remark-parse)
has used `null` instead of `undefined` for missing fields, because
`null` serializes in JSON and thus makes deep-equality tests with
fixtures stored in JSON easier,
recognizing that it is easier to have only of either `undefined` or
`null`,
given that TypeScript users might due to the existing types perform
a strict equality test against `undefined` and thus introduce bugs
such as such as:
<https://github.com/remark-embedder/core/pull/30#pullrequestreview-707605193>,
I propose to add `null` as another “missing” value to match the types
with reality.

Related to: DefinitelyTyped/DefinitelyTyped#54547.