[mdast] Add frontmatter content map
While for TypeScript block content and frontmatter content mostly the
same, it is useful to think of them separately.
Frontmatter namely exists only in top-level content, and typically only
as the head (first child).
Frontmatter does not exist in list items, block quotes, or other custom
nodes that accept block content.
The TOML node, and other potential nodes from `remark-frontmatter`,
should not be added to a `BlockContentMap`, but rather to a
`FrontmatterContentMap`.

This is documented in `syntax-tree/mdast` at the extension section for
frontmatter as `FrontmatterContent`:
<https://github.com/syntax-tree/mdast#frontmattercontent>

I also went through the comments to clarify a couple small nits.