Fix `update` function in `NodeView`
prosemirror-view@1.19.3

The update function was missing the third argument - `DecorationSource`. The argument is there in the docs - https://prosemirror.net/docs/ref/#view.NodeView.update and when you run the code the argument is also there. I gave it the `DecorationSet` type because that's its type at runtime. The docs also mention that the `DecorationSource` is implemented by `DecorationSet` (https://prosemirror.net/docs/ref/#view.DecorationSource).