Add iterator TNext type for TS 4.5
The TNext type of IIFEs is `any` by default in Typescript 4.5, not
`unknown`. That's because Typescript has different defaults for TNext
depending on whether there's a contextual type or not, and IIFEs are now
contextually typed in TS 4.5.

This is inconsistent and I'm not sure why, but it's minor enough to
document it in the tests without fixing it.