Update FfprobeFormat tags format
The current implementation of `FfprobeFormat` types `FfprobeFormat.tags` as `any[] | undefined`. This is problematic, and actually not accurate to the format in which fluent-ffmpeg returns this data.

In addition to the fact that this field seems to be where `fluent-ffmpeg` puts id3 information like artist, genre etc. the information itself is also provided as a key => value type object, rather than an array.