Make QUnit callback test context extensible in Ember QUnit
Previously, the callback to `test`, `skip`, etc. required a subtype of
of the base `TestContext` because of callback subtyping variance rules.
We didn't catch this a long time ago because the default for Ember for
unrelated reasons was having `strictFunctionTypes: false`. Enabling the
strictness setting catches this issue, though, so as downstream apps
and addons have begun using the `strict: true` setting, this showed up
as a problem.

Here, introduce a type parameter to the functions which accept
callbacks so that the variance is at the level of those functions,
rather than on the callback types themselves. This allows callers to
supply a wider type (e.g. extending `TestContext` with local values),
in line with the normal patterns in Ember testing today, where adding
more values to `this` in a test is common. This requires disabling the
DTSLint rule against single-use generics.

Additionally, the types for Ember QUnit *already* correctly handle the
new-to-TS 4.4 strictness setting `exactOptionalPropertyTypes`, for
distinguishing between `prop?: T` and `prop: T | undefined`, but the
DTSLint rules for this package had not been updated accordingly. Update
them so that tests pass!