[koa-route] Republish package to contain `types` field
The types for `koa-route` were published 3 years ago and don't contain the `types` field like new type declarations. This in turn makes it incompatible with the upcoming `node12` module resolution. The goal of this commit is simply to trigger a rebuild of the package to publish a new version with the `types` field.