rsync: Add typings for "build" static method
This change adds typings for the [build static method](https://www.npmjs.com/package/rsync#static-methods).
The build method takes an options object, where the keys are Rsync methods and the values are arguments to the methods.
I was able to accurately type this using `keyof` and `Parameters`, but to make this work efffectively I had to refactor
the overloads in the Rsync interface to use union types instead. This is because overloading and `Paramaters` do not
play well together, see https://github.com/microsoft/TypeScript/issues/32164

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>