fix: exchange params max and min of zrevrangebylex
redis official signature of zrevrangebylex is 
ZREVRANGEBYLEX key max min [LIMIT offset count]

I noticed it that in types/ioredis, the definition of zrevrangebylex misplaced the position of max and min. 
It confused me at the very beginning, I thought it might be a new implementation of official method. 
But after testing, I found it is just a simple mistake.