Allow Http2ServerRequest url to be mutable.
It has both a get and a set method defined for it. This is a backwards
compatible change, since it only compiles more programs than before.

We do not need to worry about this changing from an earlier version of
Node, since the first commit that introduced this class back around the
time of Node 8 and 9 already defined the setter.