[superagent] remove `dom` typings
At the moment, the `@types/superagent` library silently includes `dom`
typings, even in Node-only projects. This is *very* surprising.

This change sets the problematic `Response.xhr` property to `any`.

The `xhr` property is for [browser-only use][1], so it doesn't really
make sense for this browser-only property to exist on `Response`,
which currently [extends a `NodeJS.ReadableStream`][2], where you would
never expect access to DOM types.

[1]: https://github.com/visionmedia/superagent/blob/048cf185d954028b1dccde0717d2488b2284c297/src/client.js#L88
[2]: https://github.com/DefinitelyTyped/DefinitelyTyped/blob/59b04cbc8e5570ca2b8717e1cb054cfbbea01e01/types/superagent/index.d.ts#L103