rsync: Update types for build method when method doesn't take an argument
The `build` method will take an arguments object where the key is an rsync method and the value is the first paramater to that argument.
If the method doesn't take an argument (e.g. `archive`, `progress`) then the value should be "true".
Technically speaking, the value can be anything if the method doesn't take an argument, but it may as well be typed consistently here.
See the source code [here](https://github.com/mattijs/node-rsync/blob/master/rsync.js#L90)