@ember/component v4: add setComponentManager
This was actually introduced back relatively early in the 3.x series,
but it was never added -- I noticed in when looking to update it for a
deprecation which happened later in the 3.x series. As a workaround
until microsoft/DefinitelyTyped-tools#374 is merged, add a minimal set
of type definitions from `@glimmer/interfaces` to support these types.