Ember v4: remove DefaultResolver
Deprecation: https://deprecations.emberjs.com/v3.x/#toc_ember-globals-resolver

This makes the definitions in `ember-resolver` the definitive source for
the resolver types. That remains an empty stub, despite the fact that
there *is* a public API for it: the API is not documented, and this does
not change the _status quo_. Future work will add those types, and/or we
will inline the types into the package itself.