@ember/object v4: use smarter privacy technique for CPs
Most likely these will be removed entirely in short order, but as part
of the cleanup, transition the type-only fields for the get and set
types away from the older hack using many `_`s and to using private
type-only symbols. Additionally, do not export the `ComputedProperty`
*class* at all; switch to using `export type { ... }` so that only the
type side of the class gets exported for use elsewhere. This preserves
the underlying type safety, but makes it literally un-typeable, both in
terms of TS types and in terms of writing the text for them in the
source files, outside the module which defines the types.