Pin storybook-addon-jsx dependencies
@storybook/addons has nearly unusable types for adding to StoryApi --
its index signature is a union of a string and a call signature -- but
for some reason storybook-addon-jsx's intersection type works with 5.2.4
and fails with the new 5.3.22.

I don't know why, so for now I pinned the dependencies to 5.2.4
explicitly; I think the types in 6.* are significantly different so
upgrading to that will probably work around this problem.