Ember, @ember/*: `any` -> `unknown` as safe throughout
This also introduces some private type helpers to the `ember` package,
which are then used throughout, to make for safe handling of funcions,
including Ember's... very special... way of supporting passing *either*
a named value on a target object *or* a function for e.g. the runloop,
event listeners, etc.