Ember Data: introduce v4 types
Unlike with the bumps for Ember v4, this does not clean up any of the
deprecated code or improve any of the type definitions (unfortunately, I
do not have the expertise to do that correctly). It does, however, set
up the packages to keep working correctly with Ember v4, and unblock
others who *do* have that expertise to come do the cleanup work.