add(karma-image-snapshot): new definition
Declaration file for Jasmine port of 'jest-image-snapshot' for Karma
runner.
Note that DT purpose is to provide 'karma-image-snapshots' changes for
Karma configuration, DOM global object methods and for Jasmine custom
matcher, which are public APIs to be used when 'karma-image-snapshot' is
added to a project, not direct usage of 'karma-image-snapshot' module
exports (which are used by Karma dependency resolution for plugins).

'jest' related declaration have to be manually ported into DT file,
otherwise 'jest' and 'jasmine' are non-compatible in single definition,
resulting in errros which cannot be resolved on DT.

https://www.npmjs.com/package/karma-image-snapshot
https://github.com/maksimr/karma-image-snapshot#readme

Thanks!