[superagent] Fix `Blob` error
At the moment, some users' builds fail, because `Blob` isn't defined as
a global in Node.js - it sits in the [`buffer`][1] package.

This change adds an import from `buffer` to fix this issue.

Note that I think the build was incorrectly passing before because of
a [leaky interface in `stream/consumers`][2].

Note that I've had to disable the linter's `no-outside-dependencies`
because of a [known false positive on `buffer`][3].

[1]: https://nodejs.org/api/buffer.html#class-blob
[2]: https://github.com/DefinitelyTyped/DefinitelyTyped/pull/55311#discussion_r787646125
[3]: https://github.com/microsoft/dtslint/issues/315