ember-modal-dialog, ember-feature-flags: no typever for you
We had intended originally to use [the 'typever' strategy][typever] to
manage the versions of these types. However, the DefinitelyTyped tools
do not like this strategy, and while we could disable it it seems
better to keep in line with DT and to prioritize migrating these types
to the packages instead.

[typever]: https://github.com/mike-north/typever#the-versioning-strategy