ember-mocha: revert to using ember v3
We can potentially rework this in the future, but at present it still
uses the `ModuleCallbacks` type from `ember-test-helpers` in its public
interface, which means it cannot *fully* migrate to the new types from
`@ember/test-helpers`. We'll need to coordinate this with `ember-mocha`
maintainers (or just pull the types into `ember-mocha`!).