[meteor] Change user profile type to empty interface
Previously, the `profile` field on `Meteor.User` was declared as `any`.
This makes it impossible to declare a more specific type for the profile
object in your own app (it throws an error - `Subsequent property
declarations must have the same type. Property 'profile' must be of type
'any' ...`).

Instead, create a new empty `UserProfile` interface and use that as the
type of `profile`. This makes it easy for app developers to re-open the
type of `UserProfile` and add their own fields to it.

Note that this is a breaking change, as it will cause accesses to fields
under `profile` to fail, where previously they would succeed.