Update wicg-file-system-access for TS 4.6
Typescript 4.6 beta adds support for FileSystemHandle. This PR makes the
existing Definitely Typed types compatible with it.

The DOM version is missing some features.

1. FileSystemDirectoryHandle.getSystemDirectory is now gone. It was
previously deprecated, so I don't think this is a problem.

2. FileSystemHandle is now a base class, not a union. If you want to
narrow, you'll need to declare variables with type `FileSystemFileHandle |
FileSystemDirectoryHandle`. I didn't create a type alias for this yet, but
it would be a good idea.

The types here on Definitely Typed override the
`kind` property in order to allow narrowing, but the 4.6 DOM types do
not, meaning that you'll still need `@types/wicg-file-system-access` to
narrow at all. I'm going to see if I can fix that during the 4.6 beta period.