Update selenium-webdriver v4 types to reflect actual `ActionSequence` implementation/API
Selenium webdriver v4 does no longer support a button to be specified in the click shorthands. This can be observed in the JS implementation, and in the docs: 

https://github.com/SeleniumHQ/selenium/commit/b588bf52930fd4c51bbc9f9283dd0a92372b9383
https://www.selenium.dev/selenium/docs/api/javascript/module/selenium-webdriver/lib/input_exports_Actions.html

Additionally, the current types are not working anyway because `input.Button` is an enum. i.e. runtime number that
is not assignable to `string`, resulting in errors like:

```
Argument of type 'Button' is not assignable to parameter of type 'string | WebElement | undefined'.ts(2345)
```