fix: Correct types of jsforce execute arguments.
The `execute` method is intended to take an optional options argument
followed by an optional callback argument.  The first argument was
previously declared as `options: () => any | object`.  This says that
`options` is a function that returns `any` or `object` because it is
missing parens around `() => any`.

The `run` and `exec` methods are aliases for `execute`.  The `components`
and `executeAsync` methods are similar.