fix(sinon): re-use DT's fake-timers it appears @sinonjs/fake-timers stopped shipping a type.  this caused the sinon types to be stuck on the types in v7, installing that old version in node_modules.
meanwhile, @types/sinonjs_fake-timers adapted to v8 and newest sinon already uses v9.

this PR changes @types/sinon to depend on @types/sinonjs__fake-timers again, as they seem most up-to-date.

the test changes are similar to what's found in the tests for @types/sinonjs__fake-times (the casting to BrowserClock or NodeClock)