Hack around remark-validate-link warnings
It appears that `remark-validate-link` uses `hosted-git-info` to convert our GH permalink into a relative link (since it shares a prefix with our GH repo) and then warns that the relative link is dangling.  This change uses URL encoding to defeat the tool's simple prefix matching.

Filed #58598 to investigate a real solution.