Add additional typings for `Keyv.opts`
Keyv instances have a property `this.opts` (an object) which
contains both a known set of properties which are always assigned
(deserialize, namespace, serialize, store, and uri) as well as a
set of optional properties, as well as a set of arbitrary properties
which can be provided by the implementor.

This commit exposes the `Keyv.opts` property and allows the implementor
to optionally type the arbitrary options they've provided via a new,
second type parameter to the Keyv class.