react-datepicker: no `null` in max/min date
When one passes `null` as `maxDate`/`minDate` property, it is treated not as "no bound" (as `undefined` is), but as "invalid date" instead, and breaks some features such as keyboard navigation. This is very unlikely to be expected, so it'd be better to change the type for this props to `Date | undefined`.