Fix problem introduced by Typescript 4.7 prerelase
Typescript doesn't guarantee a union order and the recent release changed that order. Even if the release is reverted, this might happen again.

The culprit here is `$ExpectType` which is just not a very smart tool, since it just does basic string comparsion.

Long running solutions would either be:
- Not testing agains prereleases (thats a big decision, since all of DefinitelyTyped depends on this and it still won't rule out, this happening in the future)
- Changing tools

But since a reasonable Alternative to `$ExpectType` doesn't seem on the horizont and Typescript won't do use the favor of never changing union order, just so `$ExpectType` works, this might be our best shot for now.