Add generics for call, each, and map property keys
Previously when you would call one of the tree-walk functions (call,
each, and map) you would be asked to provide strings or numbers. So for
instance if you had an AST object like:

    { kind: "class", name: { kind: "const", value: "string" } }

you could have it recurse into the print function by calling:

    path.call(print, "name")

Unfortunately, that could also be "nam" or "namee" or "foobar" and it
wouldn't catch an error. The issue is it doesn't associate the property
key with the actual object that's being walked.

We can do better by specifying that call accepts a generic for the
property key that extends the keys of the object being walked. That
allows that property to be properly typed. There are function overrides
up to a depth of 4 properties. After that we just give up and use any so
we don't slow down compilation too much (it becomes expensive). We do a
similar thing for each and map, which have some additional complexity
because you only want to access properties that are iterable (which we
filter down to on allowable keys).