[react-tracking] useTracking generic type can be a function
When passing a function for TrackData the dependencies within TrackData,
such as `options` are actually passed the result of that function call.

Previously if you had:

```ts
useTracking<P extends Function>(getData, {
  dispatch: (data) => {
  }
});
```

The type of `data` within dispatch will be a function even though it's
actually the result of calling that function.

This PR will preserve that behavior, as P can still be any function, however if you don't pass a
function it will allow `trackingData` to be a function which returns the generic.