fix: make resolveMatch's 2nd arg type 'any'
Anything in node-tap that uses t.match() takes anything as a second
argument. The current implementation expects a string or an object of
anything, which is a problem because it disallows things like 'null' -
despite that being perfectly supported by t.match()!

This patch changes the 2nd argument to 'any', which allows consumers to
use all of the features of t.match().