[sharedb] Loosen creation types
`doc.create()` accepts an optional [type argument][1]. This type
`string` is used as a key to the object of [registered types][2], which
has both [type names][3] *and* [type URIs][4] as keys.

At the moment, the type definitions demand a type name, when a URI is
just as valid (if not actively *more* valid due to the higher degree of
specificity in a URI).

We could add the URIs on top of the existing literal type, but actually
registering arbitrary (custom, forked, etc.) types is a valid thing to
do, so this change completely loosens the type to `string`.

Note that we keep the (now unused) `OTType` definition in case any
consumers are importing it and to avoid breaking them.

[1]: https://github.com/share/sharedb/blob/92479d24e13f715fe380803d81812aa37a6e23b2/lib/client/doc.js#L882
[2]: https://github.com/share/sharedb/blob/92479d24e13f715fe380803d81812aa37a6e23b2/lib/types.js#L4
[3]: https://github.com/share/sharedb/blob/92479d24e13f715fe380803d81812aa37a6e23b2/lib/types.js#L7
[4]: https://github.com/share/sharedb/blob/92479d24e13f715fe380803d81812aa37a6e23b2/lib/types.js#L8