Update the type for the `Modal`'s `onRequestClose` callback
It receives an event object as argument, and this event might triggered by a key-press (`ESC`, for example), a blur - when the Modal loses focus (clicking outside the modal, for example) or when clicking in an element that might close it (a close button). An union of React types is provided to cover all these cases.