ArrayElement<T> should include optional array fields
I have a node (in an ast similar to estree):

export interface TypeSpecPart extends BaseNode {
  type: "TypeSpecPart";
  name: Identifier | MemberExpression;
  body?: BlockStatement;
  callspec?: MethodDefinition;
  generics?: Array<TypeSpecList>;
}

Its the only node that has a "generics" field, but since generics
is optional, its not included in ArrayElement<Node>, so that

  if (node.generics) {
    path.map(print, "generics");
  }

generates an error, saying that "generics" isn't a valid value.

This fixes the definition of ArrayElement<T> to remove null and unknown
before checking that the element is an array.