Update THREE webXR hand profile names
The hand profile names were changed in THREE.js r129 (commit 571347f). For info, the latest version of THREE.js to date is r140.
I don't know if this repository is supposed to be up to date with the latest version or this should be in a separate branch but there you go.