changed 汉语 to 中文 in README.md
The language link's name was updated to reflect the nuanced meaning of 汉语 and 中文.  @llwwbb committed on the Chinese translation should be ‘中文’ or '简体中文', and I agree with them. Although 汉语 and 中文 have nearly identical meanings, 中文 refers to the written and spoken language, while 汉语 mostly refers to the spoken language. I appreciate the comment @llwwbb.