[sinon] support missing fake timer options
At the moment, it's not possible to use some of the fake timer options,
such as [`shouldClearNativeTimers`][1], because the `sinon` types uses
its own hand-rolled [options object][2] which is missing the property.

`@types/sinon` already has a [dependency][3] on the `fake-timers`
typings, which defines its own [config object][4] with the missing
properties, so this change just uses that interface directly.

[1]: https://github.com/sinonjs/fake-timers/blob/3a77a0978eaccd73ccc87dd42204b54e2bac0f6f/src/fake-timers-src.js#L99
[2]: https://github.com/DefinitelyTyped/DefinitelyTyped/blob/658b4ded756f0c98e1a22e9592014c3c6bc4323a/types/sinon/index.d.ts#L789-L793
[3]: https://github.com/DefinitelyTyped/DefinitelyTyped/blob/658b4ded756f0c98e1a22e9592014c3c6bc4323a/types/sinon/index.d.ts#L13
[4]: https://github.com/DefinitelyTyped/DefinitelyTyped/blob/658b4ded756f0c98e1a22e9592014c3c6bc4323a/types/sinonjs__fake-timers/index.d.ts#L335-L369