@ember/routing: include back-compat shim private
In #61510, we introduced new public API import locations for these
types. However, removing the `-private` locations to import from meant
this was effectively a breaking change for existing libraries which had
little choice *other* than to use the private APIs. Providing them here
makes for a non-breaking change instead.

When we migrate these types to Ember itself, we will *not* include
these backwards-compatibility `-private` import paths, but *will*
document that it is part of the migration. A blog post announcing the
publication of these types (probably next week) will also nudge people
to start migrating.