Update node types for TS 4.9 beta (#62375)
* Update node types for TS 4.9

Add static method AbortSignal.timeout.
If this works, a later version of TS can re-add AbortSignal.abort and
node can add support for that too.

I created a ts4.8/ subdirectory for each of v14, v16, and current. I
didn't try to share types like before. With 2*3 entries I think it's
reasonably maintainable and the old way was very confusing to change.

typesVersions has a weird entry `">4.9.0-a": { "*": ["*"] }` which
works around a bug in semver parsing in Typescript. I need to ship a new
version of dtslint to allow this but I wanted to get this PR out for
people to look at.

* Update to new typesVersion format