SQLite3 Error actually extends Error
The SQLite3 Error actually extends Error versus implementing it.  I have some code that adds some common functionality to Error and expands the global type.  The implements clause is causing breakages as part of the contract is unfullfliled as SQLite3Error does not know about it.

Moving to an extends solves this problem and properply represents the relationsihp in the actual code.