node: Backport #59905 to ts4.8
Node.js types are currently out of sync between TS 4.9+ and TS <= 4.8
variants. Specifically, PR #59905 introduced many changes that were not
copied over to ts4.8. This patch is an exact copy of that PR, but
applied to ts4.8. The goal is that `@types/node` behaves the same on all
versions of TypeScript.