[node] fix: stream depends on global Blob not available in v16
Blob is only available on the global object starting in Node 18. Still,
the `stream` definitions used the global Blob. This means that
`@types/node` 16.x is currently broken. This patch adds the correct
import to resolve this problem. For more info please refer to #62654.

Long-term, IMHO we should remove `"lib": ["dom"]` from the v16 TSConfig
like it was already done for v18, but I'd much rather get this fix out
quickly and worry about that later.