Move DOM tests to root (covering ts4.9)
Remove ts4.7 directory and move one or two node-specific tests from root
to ts4.8.

This provides less test coverage; there is no exact duplicate of the
types being tested. But 4.8 and 4.9 are almost identical and 4.8 is
supposed to get every change that 4.9 does. So I think that the
DOM-specific in 4.9 and non-DOM-specific tests in 4.9 will provide
high-enough quality of coverage.