Node global interfaces redux (#62782)
* [node] First pass at separate tests under lib-dom

Made "bogus" ts-version subdirectory

* [node] Run all 4.8 tests under lib-dom

* Move DOM tests to root (covering ts4.9)

Remove ts4.7 directory and move one or two node-specific tests from root
to ts4.8.

This provides less test coverage; there is no exact duplicate of the
types being tested. But 4.8 and 4.9 are almost identical and 4.8 is
supposed to get every change that 4.9 does. So I think that the
DOM-specific in 4.9 and non-DOM-specific tests in 4.9 will provide
high-enough quality of coverage.

* Remove added whitespace

* Add missing dom-events tests

Co-authored-by: James Bromwell <943160+thw0rted@users.noreply.github.com>