Allow URL type to page path property
Thanks so much for this package! It's been super helpful trying to sort out the updated Google Analytics for our website. 

When I was trying configure GA, I found that the type URL wasn't being allowed. I had to put string. My proposal is to add URL to allow the specific type. I'm open to updating other properties to accept this change. 

TIA!