meteor: Improve specificity of publish functions
Publish functions were previously declared to return void, which meant
in practice that the return value was ignored. This changes the return
type to match Meteor's behavior, which allows either a Cursor, an array
of Cursors, or a void return (in which case the publish function is
assumed to manage records itself using this.added, this.changed, and
this.removed). Publish functions can additionally be async (i.e. return
a Promise that resolves to any of those).